Rename admin actions and reduces to users
