+++ client/reducers/index.js
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux';
 import auth from './auth';
 import documents from './documents';
-import admin from './admin';
+import users from './users';
 import search from './search';
 import paginate from './paginate';
 import user from './profile';
@@ -9,7 +9,7 @@ import user from './profile';
 export default combineReducers({
   auth,
   documents,
-  admin,
+  users,
   search,
   paginate,
   user
