+++ client/reducers/search.js
@@ -1,8 +1,8 @@
-import { SEARCH_RESULTS } from '../actions/types';
+import * as types from '../actions/types';
 
 export default function documents(state = [], action = {}) {
   switch (action.type) {
-    case SEARCH_RESULTS:
+    case types.SEARCH_RESULTS:
       return action.documentSearchResult;
     default: return state;
   }
