+++ client/reducers/users.js
@@ -1,20 +1,20 @@
-import { SET_USERS, USER_FETCHED, USER_UPDATED, USER_DELETED } from '../actions/types';
+import * as types from '../actions/types';
 
 export default function users(state = [], action = {}) {
   switch (action.type) {
-    case USER_FETCHED:
+    case types.USER_FETCHED:
       return [
         ...state,
         Object.assign({}, action.user),
       ];
-    case USER_UPDATED:
+    case types.USER_UPDATED:
       return [
         ...state.filter(user => user.id !== action.user.id),
         Object.assign({}, action.user),
       ];
-    case USER_DELETED:
+    case types.USER_DELETED:
       return state.filter(item => item.id !== action.userId);
-    case SET_USERS:
+    case types.SET_USERS:
       return action.users;
     default: return state;
   }
